+++ src/17_update_video_metadata.py
@@ -20,9 +20,10 @@ except ImportError:
 
 load_dotenv()
 
-# YouTube 영상 메타데이터 업데이트를 위해서는 youtube 스코프가 필요합니다
-# youtube.upload 스코프는 업로드만 가능하고 기존 영상 수정은 불가능합니다
-SCOPES = ['https://www.googleapis.com/auth/youtube']
+SCOPES = [
+    'https://www.googleapis.com/auth/youtube.upload',
+    'https://www.googleapis.com/auth/youtube.force-ssl'
+]
 
 
 class YouTubeMetadataUpdater: